requirements.txt
streamlit
google-generativeai
pymupdf
python-dotenv
apify-client

